requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.25.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.11.4
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
oauthlib==3.2.1
psycopg2-binary==2.9.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.1.2
zipp==3.8.0
